backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==2.30.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic-settings==2.7.0
numpy==2.3.5